dashboard/requirements.txt
streamlit==1.37.0
pandas>=2.2.2
numpy>=1.26.4
plotly>=5.22.0
requests>=2.32.3
python-dotenv>=1.0.1
pytz
duckdb>=1.1.0
plotly>=5.22.0
pytz
streamlit==1.37.0
pandas>=2.2.2
numpy>=1.26.4
python-dotenv>=1.0.1
